ref(tracing): Reduce metrics bundle size (#5122)
