+++ packages/tracing/src/browser/metrics/utils.ts
@@ -1,0 +1,26 @@
+import { Span, SpanContext } from '@sentry/types';
+
+import { Transaction } from '../../transaction';
+
+/**
+ * Checks if a given value is a valid measurement value.
+ */
+export function isMeasurementValue(value: unknown): value is number {
+  return typeof value === 'number' && isFinite(value);
+}
+
+/**
+ * Helper function to start child on transactions. This function will make sure that the transaction will
+ * use the start timestamp of the created child span if it is earlier than the transactions actual
+ * start timestamp.
+ */
+export function _startChild(transaction: Transaction, { startTimestamp, ...ctx }: SpanContext): Span {
+  if (startTimestamp && transaction.startTimestamp > startTimestamp) {
+    transaction.startTimestamp = startTimestamp;
+  }
+
+  return transaction.startChild({
+    startTimestamp,
+    ...ctx,
+  });
+}
